+++ @dao-auction/components/CurrentAuction.tsx
@@ -1,16 +1,14 @@
 import React from "react"
-import { useSigner } from "wagmi"
-import { useDaoAuctionQuery } from "@dao-auction/hooks/useDaoAuctionQuery"
 import AuctionCountdown from "./AuctionCountdown"
 import { TokenThumbnail } from "./TokenThumbnail"
 import { AuthCheck } from "../../components/elements"
+import { useActiveAuction } from "../hooks/useActiveAuction"
 
-import {
-  Auction as AuctionInterface,
-  Auction__factory as BuilderNounsAuction__factory,
-} from '@zoralabs/nouns-protocol/dist/typechain'
-import { BigNumber as EthersBN } from 'ethers'
-import { parseUnits } from '@ethersproject/units'
+/**
+ * TODO:
+ * - render bid success txHash
+ * - break ui out into atomic components
+ */
 
 export interface CurrentAuctionProps extends React.HTMLProps<HTMLDivElement> {
   /**
@@ -20,100 +18,14 @@ export interface CurrentAuctionProps extends React.HTMLProps<HTMLDivElement> {
 }
 
 export default function CurrentAuction({daoAddress, ...props}: CurrentAuctionProps) {
-  const { activeAuction } = useDaoAuctionQuery({collectionAddress: daoAddress})
-  
-  const auctionData = React.useMemo(() => {
-    const data = activeAuction?.nouns?.nounsActiveMarket
-
-    const minBidAmount = () => {
-      if (data?.highestBidPrice?.chainTokenPrice?.decimal && data?.minBidIncrementPercentage) {
-        return (data.highestBidPrice.chainTokenPrice.decimal * (data.minBidIncrementPercentage / 100)) + data.highestBidPrice.chainTokenPrice.decimal
-      } else {
-        /* @ts-ignore */
-        return data?.reservePrice?.chainTokenPrice?.decimal
-      }
-    }
-
-    return {
-      duration: data?.duration,
-      endTime: data?.endTime,
-      highestBidder: data?.highestBidder,
-      highestBidPrice: data?.highestBidPrice?.chainTokenPrice?.decimal,
-      highestBidPriceRaw: data?.highestBidPrice?.chainTokenPrice?.raw,
-      minBidIncrement: data?.minBidIncrementPercentage,
-      minBidAmount: minBidAmount(),
-      /* @ts-ignore */
-      reservePrice: data?.reservePrice?.chainTokenPrice?.raw,
-      tokenId: data?.tokenId,
-      address: data?.address,
-    }
-  }, [
-    activeAuction,
-    activeAuction?.nouns?.nounsActiveMarket?.highestBidPrice?.chainTokenPrice?.decimal
-  ])
-
-  /**
-   * Setup auction interactions
-   */
-  const [isSuccess, setIsSuccess] = React.useState(false)
-  const [isLoading, setIsLoading] = React.useState(false)
-  // const [isError, setIsError] = React.useState(false)
-  const [bidAmount, setBidAmount] = React.useState('0')
-  const [validBid, setValidBid] = React.useState(false)
-
-  const { data: signer } = useSigner()
-
-  React.useEffect(() => {
-    if (auctionData.address && signer) {
-      setBuilderNounsAuction(
-        BuilderNounsAuction__factory.connect(auctionData.address, signer)
-      )
-    }
-  }, [auctionData.address, signer])
-
-  const [BuilderNounsAuction, setBuilderNounsAuction] = React.useState<AuctionInterface>()
-
-  const handleOnUpdate = React.useCallback(
-    (value: string) => {
-      let newValue: EthersBN
-      try {
-        newValue = parseUnits(value, 18)
-        if (value >= auctionData?.minBidAmount) {
-          setValidBid(true)
-        } else {
-          setValidBid(false)
-        }
-        const bidString = newValue.toString()
-        setBidAmount(bidString)
-      } catch (e) {
-        console.error(e)
-        return
-      }
-    },
-    [setBidAmount, auctionData?.minBidAmount]
-  )
-
-  const handleOnSubmit = React.useCallback(
-    async (event: any) => {
-      if (auctionData?.tokenId) {
-        setIsLoading(true)
-        try {
-          event.preventDefault()
-          const tx = await BuilderNounsAuction?.createBid(auctionData.tokenId, {
-            value: bidAmount,
-          })
-          console.log({ tx })
-          setIsSuccess(true)
-        } catch (err: any) {
-          // setIsError(err)
-          console.error(err)
-        } finally {
-          setIsLoading(false)
-        }
-      }
-    },
-    [BuilderNounsAuction, auctionData?.tokenId, bidAmount]
-  )
+  const {
+    auctionData,
+    createBid,
+    updateBidAmount,
+    createBidSuccess,
+    createBidLoading,
+    isValidBid
+  } = useActiveAuction(daoAddress)
 
   return (
     <div className="grid grid-cols-1 lg:grid-cols-2 gap-4 max-w-[1440px]" {...props}>
@@ -129,30 +41,35 @@ export default function CurrentAuction({daoAddress, ...props}: CurrentAuctionPro
             </svg>
           </div>
         </a>
-        <div className="flex flex-row gap-10">
-          <div className="flex flex-col">
-            <span>Current Bid:</span>
-            <span>{auctionData?.highestBidPrice} ETH</span>
+        <div className="flex flex-col">
+          <div className="flex flex-row gap-10">
+            <div className="flex flex-col">
+              <span>Current Bid:</span>
+              <span>{auctionData?.highestBidPrice} ETH</span>
+            </div>
+            {auctionData?.endTime && <AuctionCountdown endTime={Number(auctionData.endTime)} />}
           </div>
-          {auctionData?.endTime && <AuctionCountdown endTime={Number(auctionData.endTime)} />}
+          <span>
+            Bidder: {auctionData?.highestBidderENS}
+          </span>
         </div>
         <AuthCheck
           connectCopy={'Connect to bid'}
           formUI={
             <div>
-              <form onSubmit={handleOnSubmit} className="flex flex-row gap-4">
+              <form onSubmit={createBid} className="flex flex-row gap-4">
                 <input
                   className="form-input px-[10px] py-[5px]"
                   type="text"
                   pattern="[0-9.]*"
                   placeholder={`${auctionData?.minBidAmount} ETH`}
-                  onChange={(event: any) => handleOnUpdate(event.target.value)}
+                  onChange={(event: any) => updateBidAmount(event.target.value)}
                 />
-                {!isLoading && !isSuccess
-                  ? <button className={`underline ${!validBid && 'pointer-events-none opacity-20'}`}>Place Bid</button>
+                {!createBidLoading && !createBidSuccess
+                  ? <button className={`underline ${!isValidBid && 'pointer-events-none opacity-20'}`}>Place Bid</button>
                   : <>
-                      {isLoading && <span>Submitting bid</span>}
-                      {isSuccess && <a href={`https://nouns.build/dao/${daoAddress}`} target="_blank" rel="noreferrer">Bid placed: view on nouns.build</a>}
+                      {createBidLoading && <span>Submitting bid</span>}
+                      {createBidSuccess && <a href={`https://nouns.build/dao/${daoAddress}`} target="_blank" rel="noreferrer">Bid placed: view on nouns.build</a>}
                     </>
                 }
               </form>
